Initial rewrite of vec without CUDA/HIP builtin vector types
