+++ include/hipSYCL/sycl/libkernel/vec.hpp
@@ -27,43 +27,154 @@
 
 #ifndef HIPSYCL_VEC_HPP
 #define HIPSYCL_VEC_HPP
-#include <cstddef>
-#include <type_traits>
-#include "hipSYCL/sycl/libkernel/backend.hpp"
-#include "hipSYCL/sycl/types.hpp"
-#include "hipSYCL/sycl/access.hpp"
+
+#include "backend.hpp"
 #include "multi_ptr.hpp"
-#include "detail/vec.hpp"
-#include "detail/vec_common.hpp"
-#include "detail/vec_swizzle.hpp"
+
+#include <cstdint>
+#include <type_traits>
+#include <utility>
 
 namespace hipsycl {
 namespace sycl {
 namespace detail {
 
+template<class T, int N>
+class vec_storage {
+public:
+  static constexpr std::size_t effective_size = (N == 3) ? 4 : N;
+  static constexpr int alignment = effective_size * N;
 
-template<class T, int N, class Function>
-HIPSYCL_UNIVERSAL_TARGET
-inline void transform_vector(vec<T,N>& v, Function f);
+  using interop_type = vec_storage<T,N>;
+  using value_type = T;
 
-template<class T, int N, class Function>
-HIPSYCL_UNIVERSAL_TARGET
-inline vec<T,N> binary_vector_operation(const vec<T,N>& a,
-                                        const vec<T,N>& b,
-                                        Function f);
+  HIPSYCL_UNIVERSAL_TARGET
+  T& operator[](int i) { return _storage[i]; }
 
-template<class T, int N, class Function>
-HIPSYCL_UNIVERSAL_TARGET
-inline vec<T,N> trinary_vector_operation(const vec<T,N>& a,
-                                         const vec<T,N>& b,
-                                         const vec<T,N>& c,
-                                         Function f);
+  HIPSYCL_UNIVERSAL_TARGET
+  const T& operator[](int i) const { return _storage[i]; }
+
+  template<int Index>
+  HIPSYCL_UNIVERSAL_TARGET
+  T& get() {
+    return _storage[Index];
+  }
+
+  template<int Index>
+  HIPSYCL_UNIVERSAL_TARGET
+  const T& get() const {
+    return _storage[Index];
+  }
+
+  template<class F>
+  HIPSYCL_UNIVERSAL_TARGET
+  void for_each(F f) {
+    for(int i = 0; i < N; ++i)
+      f(i, _storage[i]);
+  }
+
+  HIPSYCL_UNIVERSAL_TARGET
+  interop_type interop() const {
+    return *this;
+  }
+
+private:
+  alignas(alignment) T _storage [effective_size];
+};
+
+// An alternative implementation of the vec_storage concept
+// for swizzled access.
+template<class TargetStorage, int... SwizzleIndices>
+class swizzled_view_storage {
+public:
+  static constexpr std::size_t effective_size = (N == 3) ? 4 : N;
+  static constexpr int alignment = effective_size * N;
+
+  using interop_type = typename TargetStorage::interop_type;
+  using value_type = typename TargetStorage::value_type;
+
+  swizzled_view_storage(TargetStorage& s)
+  : _original_data{s} {}
+
+  HIPSYCL_UNIVERSAL_TARGET
+  value_type &operator[](int i) { 
+    return _original_data[_swizzled_indices[i]];
+  }
+
+  HIPSYCL_UNIVERSAL_TARGET
+  const value_type &operator[](int i) const {
+    return _original_data[_swizzled_indices[i]];
+  }
+
+  template<int Index>
+  HIPSYCL_UNIVERSAL_TARGET
+  value_type& get() {
+    return _original_data.template get<Index>();
+  }
+
+  template<int Index>
+  HIPSYCL_UNIVERSAL_TARGET
+  const value_type& get() const {
+    return _original_data.template get<Index>();
+  }
+
+  template<class F>
+  HIPSYCL_UNIVERSAL_TARGET
+  void for_each(F f) {
+    (f(SwizzleIndices, _original_data.template get<SwizzleIndices>()), ...);
+  }
+
+  HIPSYCL_UNIVERSAL_TARGET
+  interop_type interop() const {
+    return _original_data.interop();
+  }
+
+private:
+  static constexpr int _swizzled_indices[] = {SwizzleIndices...};
+  TargetStorage& _original_data;
+};
 
 
+template<class T>
+struct logical_vector_op_result
+{};
+
+template<> struct logical_vector_op_result<int8_t>
+{ using type = int8_t; };
+
+template<> struct logical_vector_op_result<uint8_t>
+{ using type = int8_t; };
 
 
-} // detail
+template<> struct logical_vector_op_result<int16_t>
+{ using type = int16_t; };
 
+template<> struct logical_vector_op_result<uint16_t>
+{ using type = int16_t; };
+
+template<> struct logical_vector_op_result<float>
+{ using type = int32_t; };
+
+template<> struct logical_vector_op_result<uint32_t>
+{ using type = int32_t; };
+
+template<> struct logical_vector_op_result<int32_t>
+{ using type = int32_t; };
+
+template<> struct logical_vector_op_result<double>
+{ using type = int64_t; };
+
+template<> struct logical_vector_op_result<uint64_t>
+{ using type = int64_t; };
+
+template<> struct logical_vector_op_result<int64_t>
+{ using type = int64_t; };
+
+}
+
+enum class rounding_mode {
+  automatic, rte, rtz, rtp, rtn
+};
 
 
 struct elem
@@ -94,1579 +205,552 @@ struct elem
   static constexpr int sF = 15;
 };
 
-template <typename dataT, int numElements>
-class vec
-{
-  template<class T, int N, class Function>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend void detail::transform_vector(vec<T,N>& v,
-                                       Function f);
 
-  template<class T, int N, class Function>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<T,N> detail::binary_vector_operation(const vec<T,N>& a,
-                                                  const vec<T,N>& b,
-                                                  Function f);
-
-  template<class T, int N, class Function>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<T,N> detail::trinary_vector_operation(const vec<T,N>& a,
-                                                   const vec<T,N>& b,
-                                                   const vec<T,N>& c,
-                                                   Function f);
-
-  template<class T, int N>
-  friend class vec;
-
-  HIPSYCL_UNIVERSAL_TARGET
-  explicit vec(const detail::vector_impl<dataT,numElements>& v)
-    : _impl{v}
-  {}
-
-  /// Initialization from numElements scalars
-  template<int... Indices, typename... Args>
-  HIPSYCL_UNIVERSAL_TARGET
-  void scalar_init(detail::vector_index_sequence<Indices...>, Args... args)
-  {
-    static_assert(sizeof...(args) == numElements,
-                  "Invalid number of arguments for vector "
-                  "initialization");
-
-    auto dummy_initializer = { ((_impl.template set<Indices>(args)), 0)... };
-  }
-
-  /// Initialization from single scalar
-  template<int... Indices>
-  HIPSYCL_UNIVERSAL_TARGET
-  void broadcast_scalar_init(detail::vector_index_sequence<Indices...>, dataT x)
-  {
-    auto dummy_initializer = { ((_impl.template set<Indices>(x)), 0)... };
-  }
-
-  template<int Start_index, int N, int... Indices>
-  HIPSYCL_UNIVERSAL_TARGET
-  constexpr void init_from_vec(detail::vector_index_sequence<Indices...>,
-                     const vec<dataT,N>& other)
-  {
-    auto dummy_initializer =
-    {
-      ((_impl.template set<Start_index + Indices>(
-                                   other._impl.template get<Indices>())), 0)...
-    };
-  }
-
-  template<int Start_index, int N>
-  HIPSYCL_UNIVERSAL_TARGET
-  constexpr void init_from_vec(const vec<dataT,N>& other)
-  {
-    init_from_vec<Start_index>(typename detail::vector_impl<dataT,N>::indices(),
-                               other);
-  }
-
-  template<int Position>
-  HIPSYCL_UNIVERSAL_TARGET
-  constexpr void init_from_argument()
-  {
-    static_assert (Position == numElements, "Invalid number of vector "
-                                            "constructor arguments.");
-  }
-
-  template<int Position, typename... Args>
-  HIPSYCL_UNIVERSAL_TARGET
-  constexpr void init_from_argument(dataT scalar,
-                                    const Args&... args)
-  {
-    _impl.template set<Position>(scalar);
-    init_from_argument<Position+1>(args...);
-  }
-
-  template<int Position, int N, typename... Args>
-  HIPSYCL_UNIVERSAL_TARGET
-  constexpr void init_from_argument(const vec<dataT,N>& v,
-                                    const Args&... args)
-  {
-    init_from_vec<Position>(v);
-    init_from_argument<Position + N>(args...);
-  }
-
-  template<int... swizzleIndices>
-  HIPSYCL_UNIVERSAL_TARGET
-  auto swizzle_index_sequence(detail::vector_index_sequence<swizzleIndices...>) const
-  {
-    return swizzle<swizzleIndices...>();
-  }
+template<typename T, int N, class VectorStorage = detail::vec_storage<T,N>>
+class vec {
 public:
-  static_assert(numElements == 1 ||
-                numElements == 2 ||
-                numElements == 3 ||
-                numElements == 4 ||
-                numElements == 8 ||
-                numElements == 16,
-                "Invalid number of vector elements. Allowed values: "
-                "1,2,3,4,8,16");
+  static_assert(N == 1 || N == 2 || N == 3 || N == 4 || N == 8 || N == 16,
+                "Invalid number of vec elements");
 
-  using element_type = dataT;
+  using element_type = T;
 
-#ifdef SYCL_DEVICE_ONLY
-  using vector_t = vec<dataT, numElements>;
-#endif
-  HIPSYCL_UNIVERSAL_TARGET
-  vec()
-    : _impl {}
-  {}
-
-  template<int Original_vector_size,
-           int... Swizzle_indices>
-  HIPSYCL_UNIVERSAL_TARGET
-  vec(const detail::vec_swizzle<dataT, Original_vector_size, Swizzle_indices...>&
-      swizzled_vec)
-    : _impl{swizzled_vec.swizzled_access().get()}
-  {}
+  using vector_t = typename VectorStorage::interop_type;
 
   HIPSYCL_UNIVERSAL_TARGET
-  explicit vec(const dataT &arg)
-  {
-    broadcast_scalar_init(typename detail::vector_impl<dataT,numElements>::indices(),
-                          arg);
+  vec() {
+    for(int i = 0; i < N; ++i)
+      _data[i] = T{};
   }
 
-  template <typename... argTN>
   HIPSYCL_UNIVERSAL_TARGET
-  vec(const argTN&... args)
-  {
-    init_from_argument<0>(args...);
+  vec(VectorStorage v)
+  : _data{v} {}
+
+  HIPSYCL_UNIVERSAL_TARGET
+  explicit vec(const T& value) {
+    for(int i = 0; i < N; ++i)
+      _data[i] = value;
+  }
+
+  template<typename... Args>
+  HIPSYCL_UNIVERSAL_TARGET
+  vec(const Args&... args) {
+    static_assert((count_num_elements(args) + ...) == N,
+                  "Argument mismatch with vector size");
+
+    int current_init_index = 0;
+    (partial_initialization(current_init_index, args), ...);
+  }
+
+  template<class Storage>
+  HIPSYCL_UNIVERSAL_TARGET
+  vec(const vec<T, N, Storage>& other) {
+    for(int i = 0; i < N; ++i)
+      _data[i] = other._data[i];
+  }
+
+  HIPSYCL_UNIVERSAL_TARGET
+  vec(vector_t v)
+  : _data{v} {}
+
+  HIPSYCL_UNIVERSAL_TARGET
+  operator vector_t() const {
+    return _data.interop();
+  }
+
+  template <int Dim = N, std::enable_if_t<Dim == 1, bool> = true>
+  HIPSYCL_UNIVERSAL_TARGET
+  operator T() const { return _data.template get<0>(); }
+
+  HIPSYCL_UNIVERSAL_TARGET
+  static constexpr int get_count() { return N; }
+
+  HIPSYCL_UNIVERSAL_TARGET
+  static constexpr std::size_t get_size() { return sizeof(VectorStorage); }
+
+  template <typename ConvertT, rounding_mode RM = rounding_mode::automatic>
+  HIPSYCL_UNIVERSAL_TARGET
+  vec<ConvertT, N> convert() const {
+
+    vec<ConvertT, N> result;
+
+    for(int i = 0; i < N; ++i) {
+      // TODO: Take rounding mode into account
+      result._data[i] = static_cast<ConvertT>(_data[i]);
+    }
+
+    return result;
+  }
+
+  template<typename AsT, int OtherN>
+  HIPSYCL_UNIVERSAL_TARGET
+  vec<AsT, OtherN> as() const {
+    static_assert(sizeof(vec<AsT, OtherN>) == sizeof(vec<T,N>),
+                  "Reinterpreted vector must have same size");
+    static_assert(std::is_same_v<VectorStorage, detail::vec_storage<T, N>>,
+                  "Reinterpreting swizzled vectors directly is not supported");
+
+    vec<AsT, N> result;
+    
+    AsT* in_ptr = reinterpret_cast<AsT*>(&_data[0]);
+    for(int i = 0; i < OtherN; ++i)
+      result._data[i] = in_ptr[i];
+
+    return result;
+  }
+
+  template<int... SwizzleIndices>
+  HIPSYCL_UNIVERSAL_TARGET  
+  auto swizzle() const {
+
+    using swizzle_view_type =
+        detail::swizzled_view_storage<VectorStorage, SwizzleIndices...>;
+
+    VectorStorage& data = const_cast<VectorStorage&>(_data);
+
+    return vec<T, sizeof...(SwizzleIndices), swizzle_view_type>{
+        swizzle_view_type{data}};
+  }
+
+  template<int Dim = N, std::enable_if_t<(Dim > 1), bool> = true>
+  HIPSYCL_UNIVERSAL_TARGET
+  auto lo() const {
+    if constexpr(Dim == 2) {
+      return swizzle<0>();
+    } else if constexpr(Dim == 3 || Dim == 4) {
+      return swizzle<0,1>();
+    } else if constexpr(Dim == 8) {
+      return swizzle<0,1,2,3>();
+    } else if constexpr(Dim == 16) {
+      return swizzle<0,1,2,3,4,5,6,7>();
+    } else {
+      // Should never reach
+      return *this;
+    }
+  }
+
+  template<int Dim = N, std::enable_if_t<(Dim > 1), bool> = true>
+  HIPSYCL_UNIVERSAL_TARGET
+  auto hi() const {
+    if constexpr(Dim == 2) {
+      return swizzle<1>();
+    } else if constexpr(Dim == 3 || Dim == 4) {
+      return swizzle<2,3>();
+    } else if constexpr(Dim == 8) {
+      return swizzle<4,5,6,7>();
+    } else if constexpr(Dim == 16) {
+      return swizzle<8,9,10,11,12,13,14,15>();
+    } else {
+      // Should never reach
+      return *this;
+    }
+  }
+
+  template<int Dim = N, std::enable_if_t<(Dim > 1), bool> = true>
+  HIPSYCL_UNIVERSAL_TARGET
+  auto odd() const {
+    if constexpr(Dim == 2) {
+      return swizzle<1>();
+    } else if constexpr(Dim == 3 || Dim == 4) {
+      return swizzle<1,3>();
+    } else if constexpr(Dim == 8) {
+      return swizzle<1,3,5,7>();
+    } else if constexpr(Dim == 16) {
+      return swizzle<1,3,5,7,9,11,13,15>();
+    } else {
+      // Should never reach
+      return *this;
+    }
   }
 
 
-  vec(const vec<dataT, numElements> &rhs) = default;
-
-  // OpenCL interop is unsupported
-//#ifdef SYCL_DEVICE_ONLY
-  //vec(vector_t openclVector);
-  //operator vector_t() const;
-//#endif
-
-  // Available only when: numElements == 1
-  template<int N = numElements,
-           std::enable_if_t<N == 1>* = nullptr>
+  template<int Dim = N, std::enable_if_t<(Dim > 1), bool> = true>
   HIPSYCL_UNIVERSAL_TARGET
-  operator dataT() const
-  { return _impl.template get<0>(); }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  static constexpr int get_count()
-  { return numElements; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  static constexpr size_t get_size()
-  { return numElements * sizeof (dataT); }
-
-  // ToDo
-  template <typename convertT, rounding_mode roundingMode>
-  vec<convertT, numElements> convert() const;
-
-  // ToDo
-  template <typename asT>
-  asT as() const;
-
-#define HIPSYCL_DEFINE_VECTOR_ACCESS_IF(condition, name, id) \
-  template<int N = numElements, \
-           std::enable_if_t<(id < N) && (condition)>* = nullptr> \
-  HIPSYCL_UNIVERSAL_TARGET \
-  dataT& name() \
-  { return _impl.template get<id>(); } \
-  \
-  template<int N = numElements, \
-           std::enable_if_t<(id < N) && (condition)>* = nullptr> \
-  HIPSYCL_UNIVERSAL_TARGET \
-  dataT name() const \
-  { return _impl.template get<id>(); }
-
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 4, x, 0)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 4, y, 1)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 4, z, 2)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 4, w, 3)
-
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements == 4, r, 0)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements == 4, g, 1)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements == 4, b, 2)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements == 4, a, 3)
-
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, s0, 0)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, s1, 1)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, s2, 2)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, s3, 3)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, s4, 4)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, s5, 5)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, s6, 6)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, s7, 7)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, s8, 8)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, s9, 9)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, sA, 10)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, sB, 11)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, sC, 12)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, sD, 13)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, sE, 14)
-  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(numElements <= 16, sF, 15)
-
-
-  template<int... swizzleIndexes>
-  HIPSYCL_UNIVERSAL_TARGET
-  auto swizzle() const
-  {
-    return detail::vec_swizzle<dataT, numElements, swizzleIndexes...>{
-      const_cast<detail::vector_impl<dataT,numElements>&>(_impl)
-    };
+  auto even() const {
+    if constexpr(Dim == 2) {
+      return swizzle<0>();
+    } else if constexpr(Dim == 3 || Dim == 4) {
+      return swizzle<0,2>();
+    } else if constexpr(Dim == 8) {
+      return swizzle<0,2,4,6>();
+    } else if constexpr(Dim == 16) {
+      return swizzle<0,2,4,6,8,10,12,14>();
+    } else {
+      // Should never reach
+      return *this;
+    }
   }
 
-  template<int N = numElements,
-           std::enable_if_t<(N > 1)>* = nullptr>
+
   HIPSYCL_UNIVERSAL_TARGET
-  auto lo() const
-  {
-    return swizzle_index_sequence(
-          typename detail::vector_impl<dataT, numElements>::lo_indices());
+  T& operator[](int index) {
+    return _data[index];
   }
 
-  template<int N = numElements,
-           std::enable_if_t<(N > 1)>* = nullptr>
   HIPSYCL_UNIVERSAL_TARGET
-  auto hi() const
-  {
-    return swizzle_index_sequence(
-          typename detail::vector_impl<dataT, numElements>::hi_indices());
+  const T& operator[](int index) const {
+    return _data[index];
   }
 
-  template<int N = numElements,
-           std::enable_if_t<(N > 1)>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  auto even() const
-  {
-    return swizzle_index_sequence(
-          typename detail::vector_impl<dataT, numElements>::even_indices());
+  vec& operator=(const vec& rhs) {
+    for(int i = 0; i < N; ++i)
+      _data[i] = rhs._data[i];
+    return *this;
   }
 
-  template<int N = numElements,
-           std::enable_if_t<(N > 1)>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  auto odd() const
-  {
-    return swizzle_index_sequence(
-          typename detail::vector_impl<dataT, numElements>::odd_indices());
+  vec& operator=(const T& rhs) {
+    for(int i = 0; i < N; ++i)
+      _data[i] = rhs;
+    return *this;
   }
 
-#ifdef SYCL_SIMPLE_SWIZZLES
-#define HIPSYCL_DEFINE_VECTOR_SWIZZLE2(name, N, i0, i1) \
-  template<int n = numElements, \
-           std::enable_if_t<(n >= N && n <= 4)>* = nullptr> \
+#define HIPSYCL_DEFINE_VECTOR_ACCESS_IF(condition, name, id)                   \
+  template <int Dim = N,                                                       \
+            std::enable_if_t<(id < N) && (condition), bool> = true>            \
+  HIPSYCL_UNIVERSAL_TARGET T &name() {                                         \
+    return _data.template get<id>();                                           \
+  }                                                                            \
+                                                                               \
+  template <int Dim = N,                                                       \
+            std::enable_if_t<(id < N) && (condition), bool> = true>            \
+  HIPSYCL_UNIVERSAL_TARGET T name() const {                                    \
+    return _data.template get<id>();                                           \
+  }
+
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 4, x, 0)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 4, y, 1)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 4, z, 2)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 4, w, 3)
+
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N == 4, r, 0)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N == 4, g, 1)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N == 4, b, 2)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N == 4, a, 3)
+
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, s0, 0)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, s1, 1)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, s2, 2)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, s3, 3)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, s4, 4)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, s5, 5)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, s6, 6)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, s7, 7)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, s8, 8)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, s9, 9)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, sA, 10)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, sB, 11)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, sC, 12)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, sD, 13)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, sE, 14)
+  HIPSYCL_DEFINE_VECTOR_ACCESS_IF(N <= 16, sF, 15)
+
+#define HIPSYCL_DEFINE_VECTOR_SWIZZLE2(name, Dim, i0, i1) \
+  template<int n = N, \
+           std::enable_if_t<(n >= Dim && n <= 4), bool> = true> \
   HIPSYCL_UNIVERSAL_TARGET \
   auto name() const \
   {return swizzle<i0,i1>(); }
 
-#define HIPSYCL_DEFINE_VECTOR_SWIZZLE3(name, N, i0, i1, i2) \
-  template<int n = numElements, \
-           std::enable_if_t<(n >= N && n <= 4)>* = nullptr> \
+#define HIPSYCL_DEFINE_VECTOR_SWIZZLE3(name, Dim, i0, i1, i2) \
+  template<int n = N, \
+           std::enable_if_t<(n >= Dim && n <= 4), bool> = true> \
   HIPSYCL_UNIVERSAL_TARGET \
   auto name() const \
   {return swizzle<i0,i1,i2>(); }
 
 #define HIPSYCL_DEFINE_VECTOR_SWIZZLE4(name, i0, i1, i2, i3) \
-  template<int n = numElements, \
+  template<int n = N, \
            std::enable_if_t<(n == 4)>* = nullptr> \
   HIPSYCL_UNIVERSAL_TARGET \
   auto name() const \
   {return swizzle<i0,i1,i2,i3>(); }
 
-
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(xx, 2, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(yx, 2, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(xy, 2, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(yy, 2, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(xz, 3, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(yz, 3, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(zy, 3, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(zx, 3, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(zz, 3, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(xw, 4, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(yw, 4, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(zw, 4, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(ww, 4, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(wx, 4, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(wy, 4, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE2(wz, 4, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xxx, 3, 0, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(yxx, 3, 1, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zxx, 3, 2, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xyx, 3, 0, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(yyx, 3, 1, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zyx, 3, 2, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xzx, 3, 0, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(yzx, 3, 1, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zzx, 3, 2, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xxy, 3, 0, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(yxy, 3, 1, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zxy, 3, 2, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xyy, 3, 0, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(yyy, 3, 1, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zyy, 3, 2, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xzy, 3, 0, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(yzy, 3, 1, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zzy, 3, 2, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xxz, 3, 0, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(yxz, 3, 1, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zxz, 3, 2, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xyz, 3, 0, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(yyz, 3, 1, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zyz, 3, 2, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xzz, 3, 0, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(yzz, 3, 1, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zzz, 3, 2, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xxw, 4, 0, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(yxw, 4, 1, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zxw, 4, 2, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xyw, 4, 0, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(yyw, 4, 1, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zyw, 4, 2, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xzw, 4, 0, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(yzw, 4, 1, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zzw, 4, 2, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xwx, 4, 0, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(ywx, 4, 1, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zwx, 4, 2, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xwy, 4, 0, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(ywy, 4, 1, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zwy, 4, 2, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xwz, 4, 0, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(ywz, 4, 1, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zwz, 4, 2, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(wxx, 4, 3, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(wyx, 4, 3, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(wzx, 4, 3, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(wxy, 4, 3, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(wyy, 4, 3, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(wzy, 4, 3, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(wxz, 4, 3, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(wyz, 4, 3, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(wzz, 4, 3, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(xww, 4, 0, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(yww, 4, 1, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(zww, 4, 2, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(wwx, 4, 3, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(wwy, 4, 3, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(wwz, 4, 3, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE3(www, 4, 3, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxxx, 0, 0, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrrr, 0, 0, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxxx, 1, 0, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grrr, 1, 0, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxxx, 2, 0, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brrr, 2, 0, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxxx, 3, 0, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(arrr, 3, 0, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xyxx, 0, 1, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rgrr, 0, 1, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yyxx, 1, 1, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ggrr, 1, 1, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zyxx, 2, 1, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bgrr, 2, 1, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wyxx, 3, 1, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(agrr, 3, 1, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzxx, 0, 2, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbrr, 0, 2, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzxx, 1, 2, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbrr, 1, 2, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzxx, 2, 2, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbrr, 2, 2, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzxx, 3, 2, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abrr, 3, 2, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwxx, 0, 3, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rarr, 0, 3, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywxx, 1, 3, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(garr, 1, 3, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwxx, 2, 3, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(barr, 2, 3, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwxx, 3, 3, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aarr, 3, 3, 0, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxyx, 0, 0, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrgr, 0, 0, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxyx, 1, 0, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grgr, 1, 0, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxyx, 2, 0, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brgr, 2, 0, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxyx, 3, 0, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(argr, 3, 0, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xyyx, 0, 1, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rggr, 0, 1, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yyyx, 1, 1, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gggr, 1, 1, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zyyx, 2, 1, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bggr, 2, 1, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wyyx, 3, 1, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aggr, 3, 1, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzyx, 0, 2, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbgr, 0, 2, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzyx, 1, 2, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbgr, 1, 2, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzyx, 2, 2, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbgr, 2, 2, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzyx, 3, 2, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abgr, 3, 2, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwyx, 0, 3, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ragr, 0, 3, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywyx, 1, 3, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gagr, 1, 3, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwyx, 2, 3, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bagr, 2, 3, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwyx, 3, 3, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aagr, 3, 3, 1, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxzx, 0, 0, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrbr, 0, 0, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxzx, 1, 0, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grbr, 1, 0, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxzx, 2, 0, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brbr, 2, 0, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxzx, 3, 0, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(arbr, 3, 0, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xyzx, 0, 1, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rgbr, 0, 1, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yyzx, 1, 1, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ggbr, 1, 1, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zyzx, 2, 1, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bgbr, 2, 1, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wyzx, 3, 1, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(agbr, 3, 1, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzzx, 0, 2, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbbr, 0, 2, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzzx, 1, 2, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbbr, 1, 2, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzzx, 2, 2, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbbr, 2, 2, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzzx, 3, 2, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abbr, 3, 2, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwzx, 0, 3, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rabr, 0, 3, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywzx, 1, 3, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gabr, 1, 3, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwzx, 2, 3, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(babr, 2, 3, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwzx, 3, 3, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aabr, 3, 3, 2, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxwx, 0, 0, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrar, 0, 0, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxwx, 1, 0, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grar, 1, 0, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxwx, 2, 0, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brar, 2, 0, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxwx, 3, 0, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(arar, 3, 0, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xywx, 0, 1, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rgar, 0, 1, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yywx, 1, 1, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ggar, 1, 1, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zywx, 2, 1, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bgar, 2, 1, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wywx, 3, 1, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(agar, 3, 1, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzwx, 0, 2, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbar, 0, 2, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzwx, 1, 2, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbar, 1, 2, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzwx, 2, 2, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbar, 2, 2, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzwx, 3, 2, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abar, 3, 2, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwwx, 0, 3, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(raar, 0, 3, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywwx, 1, 3, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gaar, 1, 3, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwwx, 2, 3, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(baar, 2, 3, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwwx, 3, 3, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aaar, 3, 3, 3, 0)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxxy, 0, 0, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrrg, 0, 0, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxxy, 1, 0, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grrg, 1, 0, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxxy, 2, 0, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brrg, 2, 0, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxxy, 3, 0, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(arrg, 3, 0, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xyxy, 0, 1, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rgrg, 0, 1, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yyxy, 1, 1, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ggrg, 1, 1, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zyxy, 2, 1, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bgrg, 2, 1, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wyxy, 3, 1, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(agrg, 3, 1, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzxy, 0, 2, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbrg, 0, 2, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzxy, 1, 2, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbrg, 1, 2, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzxy, 2, 2, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbrg, 2, 2, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzxy, 3, 2, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abrg, 3, 2, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwxy, 0, 3, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rarg, 0, 3, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywxy, 1, 3, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(garg, 1, 3, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwxy, 2, 3, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(barg, 2, 3, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwxy, 3, 3, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aarg, 3, 3, 0, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxyy, 0, 0, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrgg, 0, 0, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxyy, 1, 0, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grgg, 1, 0, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxyy, 2, 0, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brgg, 2, 0, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxyy, 3, 0, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(argg, 3, 0, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xyyy, 0, 1, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rggg, 0, 1, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yyyy, 1, 1, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gggg, 1, 1, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zyyy, 2, 1, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bggg, 2, 1, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wyyy, 3, 1, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aggg, 3, 1, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzyy, 0, 2, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbgg, 0, 2, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzyy, 1, 2, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbgg, 1, 2, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzyy, 2, 2, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbgg, 2, 2, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzyy, 3, 2, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abgg, 3, 2, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwyy, 0, 3, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ragg, 0, 3, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywyy, 1, 3, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gagg, 1, 3, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwyy, 2, 3, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bagg, 2, 3, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwyy, 3, 3, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aagg, 3, 3, 1, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxzy, 0, 0, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrbg, 0, 0, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxzy, 1, 0, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grbg, 1, 0, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxzy, 2, 0, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brbg, 2, 0, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxzy, 3, 0, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(arbg, 3, 0, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xyzy, 0, 1, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rgbg, 0, 1, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yyzy, 1, 1, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ggbg, 1, 1, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zyzy, 2, 1, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bgbg, 2, 1, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wyzy, 3, 1, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(agbg, 3, 1, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzzy, 0, 2, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbbg, 0, 2, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzzy, 1, 2, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbbg, 1, 2, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzzy, 2, 2, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbbg, 2, 2, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzzy, 3, 2, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abbg, 3, 2, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwzy, 0, 3, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rabg, 0, 3, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywzy, 1, 3, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gabg, 1, 3, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwzy, 2, 3, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(babg, 2, 3, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwzy, 3, 3, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aabg, 3, 3, 2, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxwy, 0, 0, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrag, 0, 0, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxwy, 1, 0, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grag, 1, 0, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxwy, 2, 0, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brag, 2, 0, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxwy, 3, 0, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(arag, 3, 0, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xywy, 0, 1, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rgag, 0, 1, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yywy, 1, 1, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ggag, 1, 1, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zywy, 2, 1, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bgag, 2, 1, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wywy, 3, 1, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(agag, 3, 1, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzwy, 0, 2, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbag, 0, 2, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzwy, 1, 2, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbag, 1, 2, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzwy, 2, 2, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbag, 2, 2, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzwy, 3, 2, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abag, 3, 2, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwwy, 0, 3, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(raag, 0, 3, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywwy, 1, 3, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gaag, 1, 3, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwwy, 2, 3, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(baag, 2, 3, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwwy, 3, 3, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aaag, 3, 3, 3, 1)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxxz, 0, 0, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrrb, 0, 0, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxxz, 1, 0, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grrb, 1, 0, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxxz, 2, 0, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brrb, 2, 0, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxxz, 3, 0, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(arrb, 3, 0, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xyxz, 0, 1, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rgrb, 0, 1, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yyxz, 1, 1, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ggrb, 1, 1, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zyxz, 2, 1, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bgrb, 2, 1, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wyxz, 3, 1, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(agrb, 3, 1, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzxz, 0, 2, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbrb, 0, 2, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzxz, 1, 2, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbrb, 1, 2, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzxz, 2, 2, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbrb, 2, 2, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzxz, 3, 2, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abrb, 3, 2, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwxz, 0, 3, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rarb, 0, 3, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywxz, 1, 3, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(garb, 1, 3, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwxz, 2, 3, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(barb, 2, 3, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwxz, 3, 3, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aarb, 3, 3, 0, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxyz, 0, 0, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrgb, 0, 0, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxyz, 1, 0, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grgb, 1, 0, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxyz, 2, 0, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brgb, 2, 0, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxyz, 3, 0, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(argb, 3, 0, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xyyz, 0, 1, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rggb, 0, 1, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yyyz, 1, 1, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gggb, 1, 1, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zyyz, 2, 1, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bggb, 2, 1, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wyyz, 3, 1, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aggb, 3, 1, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzyz, 0, 2, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbgb, 0, 2, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzyz, 1, 2, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbgb, 1, 2, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzyz, 2, 2, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbgb, 2, 2, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzyz, 3, 2, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abgb, 3, 2, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwyz, 0, 3, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ragb, 0, 3, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywyz, 1, 3, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gagb, 1, 3, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwyz, 2, 3, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bagb, 2, 3, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwyz, 3, 3, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aagb, 3, 3, 1, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxzz, 0, 0, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrbb, 0, 0, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxzz, 1, 0, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grbb, 1, 0, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxzz, 2, 0, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brbb, 2, 0, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxzz, 3, 0, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(arbb, 3, 0, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xyzz, 0, 1, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rgbb, 0, 1, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yyzz, 1, 1, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ggbb, 1, 1, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zyzz, 2, 1, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bgbb, 2, 1, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wyzz, 3, 1, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(agbb, 3, 1, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzzz, 0, 2, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbbb, 0, 2, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzzz, 1, 2, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbbb, 1, 2, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzzz, 2, 2, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbbb, 2, 2, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzzz, 3, 2, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abbb, 3, 2, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwzz, 0, 3, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rabb, 0, 3, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywzz, 1, 3, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gabb, 1, 3, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwzz, 2, 3, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(babb, 2, 3, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwzz, 3, 3, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aabb, 3, 3, 2, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxwz, 0, 0, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrab, 0, 0, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxwz, 1, 0, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grab, 1, 0, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxwz, 2, 0, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brab, 2, 0, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxwz, 3, 0, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(arab, 3, 0, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xywz, 0, 1, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rgab, 0, 1, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yywz, 1, 1, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ggab, 1, 1, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zywz, 2, 1, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bgab, 2, 1, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wywz, 3, 1, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(agab, 3, 1, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzwz, 0, 2, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbab, 0, 2, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzwz, 1, 2, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbab, 1, 2, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzwz, 2, 2, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbab, 2, 2, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzwz, 3, 2, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abab, 3, 2, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwwz, 0, 3, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(raab, 0, 3, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywwz, 1, 3, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gaab, 1, 3, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwwz, 2, 3, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(baab, 2, 3, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwwz, 3, 3, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aaab, 3, 3, 3, 2)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxxw, 0, 0, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrra, 0, 0, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxxw, 1, 0, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grra, 1, 0, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxxw, 2, 0, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brra, 2, 0, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxxw, 3, 0, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(arra, 3, 0, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xyxw, 0, 1, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rgra, 0, 1, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yyxw, 1, 1, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ggra, 1, 1, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zyxw, 2, 1, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bgra, 2, 1, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wyxw, 3, 1, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(agra, 3, 1, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzxw, 0, 2, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbra, 0, 2, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzxw, 1, 2, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbra, 1, 2, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzxw, 2, 2, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbra, 2, 2, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzxw, 3, 2, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abra, 3, 2, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwxw, 0, 3, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rara, 0, 3, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywxw, 1, 3, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gara, 1, 3, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwxw, 2, 3, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bara, 2, 3, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwxw, 3, 3, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aara, 3, 3, 0, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxyw, 0, 0, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrga, 0, 0, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxyw, 1, 0, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grga, 1, 0, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxyw, 2, 0, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brga, 2, 0, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxyw, 3, 0, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(arga, 3, 0, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xyyw, 0, 1, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rgga, 0, 1, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yyyw, 1, 1, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ggga, 1, 1, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zyyw, 2, 1, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bgga, 2, 1, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wyyw, 3, 1, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(agga, 3, 1, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzyw, 0, 2, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbga, 0, 2, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzyw, 1, 2, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbga, 1, 2, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzyw, 2, 2, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbga, 2, 2, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzyw, 3, 2, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abga, 3, 2, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwyw, 0, 3, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(raga, 0, 3, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywyw, 1, 3, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gaga, 1, 3, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwyw, 2, 3, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(baga, 2, 3, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwyw, 3, 3, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aaga, 3, 3, 1, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxzw, 0, 0, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rrba, 0, 0, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxzw, 1, 0, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(grba, 1, 0, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxzw, 2, 0, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(brba, 2, 0, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxzw, 3, 0, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(arba, 3, 0, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xyzw, 0, 1, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rgba, 0, 1, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yyzw, 1, 1, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ggba, 1, 1, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zyzw, 2, 1, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bgba, 2, 1, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wyzw, 3, 1, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(agba, 3, 1, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzzw, 0, 2, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbba, 0, 2, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzzw, 1, 2, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbba, 1, 2, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzzw, 2, 2, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbba, 2, 2, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzzw, 3, 2, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abba, 3, 2, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwzw, 0, 3, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(raba, 0, 3, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywzw, 1, 3, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gaba, 1, 3, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwzw, 2, 3, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(baba, 2, 3, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwzw, 3, 3, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aaba, 3, 3, 2, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xxww, 0, 0, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rraa, 0, 0, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yxww, 1, 0, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(graa, 1, 0, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zxww, 2, 0, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(braa, 2, 0, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wxww, 3, 0, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(araa, 3, 0, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xyww, 0, 1, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rgaa, 0, 1, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yyww, 1, 1, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ggaa, 1, 1, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zyww, 2, 1, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bgaa, 2, 1, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wyww, 3, 1, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(agaa, 3, 1, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xzww, 0, 2, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(rbaa, 0, 2, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(yzww, 1, 2, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gbaa, 1, 2, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zzww, 2, 2, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(bbaa, 2, 2, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wzww, 3, 2, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(abaa, 3, 2, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(xwww, 0, 3, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(raaa, 0, 3, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(ywww, 1, 3, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(gaaa, 1, 3, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(zwww, 2, 3, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(baaa, 2, 3, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(wwww, 3, 3, 3, 3)
-  HIPSYCL_DEFINE_VECTOR_SWIZZLE4(aaaa, 3, 3, 3, 3)
+#ifdef SYCL_SIMPLE_SWIZZLES
+ #include "detail/vec_simple_swizzles.hpp"
 #endif
 
-
   // ToDo: load and store member functions
-  template <access::address_space addressSpace>
+  // ToDo: These functions have changed signatures with SYCL 2020
+  template <access::address_space AddressSpace>
   HIPSYCL_UNIVERSAL_TARGET
-  void load(size_t offset, multi_ptr<dataT, addressSpace> ptr);
-  template <access::address_space addressSpace>
-  HIPSYCL_UNIVERSAL_TARGET
-  void store(size_t offset, multi_ptr<dataT, addressSpace> ptr) const;
+  void load(size_t offset, multi_ptr<T, AddressSpace> ptr);
 
-  // OP is: +, -, *, /, %
-  /* When OP is % available only when: dataT != cl_float && dataT != cl_double
-&& dataT != cl_half. */
+  template <access::address_space AddressSpace>
   HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> operator+(const vec<dataT, numElements> &lhs, 
-                                           const vec<dataT, numElements> &rhs) 
-  { return vec<dataT,numElements>{lhs._impl + rhs._impl}; }
+  void store(size_t offset, multi_ptr<T, AddressSpace> ptr) const;
 
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> operator-(const vec<dataT, numElements> &lhs,
-                                           const vec<dataT, numElements> &rhs)
-  { return vec<dataT,numElements>{lhs._impl - rhs._impl}; }
+#define HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_VEC(op, t)                            \
+  HIPSYCL_UNIVERSAL_TARGET                                                     \
+  friend vec<t, N> operator op(const vec<t, N> &lhs, const vec<t, N> &rhs) {   \
+    vec<t, N> result;                                                          \
+    for (int i = 0; i < N; ++i) {                                              \
+      result._data[i] = lhs._data[i] op rhs._data[i];                          \
+    }                                                                          \
+    return result;                                                             \
+  }
 
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> operator*(const vec<dataT, numElements> &lhs,
-                                           const vec<dataT, numElements> &rhs)
-  { return vec<dataT,numElements>{lhs._impl * rhs._impl}; }
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_VEC(%, t)
 
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> operator/(const vec<dataT, numElements> &lhs,
-                                           const vec<dataT, numElements> &rhs)
-  { return vec<dataT,numElements>{lhs._impl / rhs._impl}; }
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_VEC(+, T)
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_VEC(-, T)
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_VEC(*, T)
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_VEC(/, T)
 
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> operator%(const vec<dataT, numElements> &lhs,
-                                           const vec<dataT, numElements> &rhs)
-  { return vec<dataT,numElements>{lhs._impl % rhs._impl}; }
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_VEC(&, t)
+  
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_VEC(|, t)
 
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> operator+(const vec<dataT, numElements> &lhs,
-                                           const dataT &rhs)
-  { return vec<dataT,numElements>{lhs._impl + rhs}; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> operator-(const vec<dataT, numElements> &lhs,
-                                           const dataT &rhs)
-  { return vec<dataT,numElements>{lhs._impl - rhs}; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> operator*(const vec<dataT, numElements> &lhs,
-                                           const dataT &rhs)
-  { return vec<dataT,numElements>{lhs._impl * rhs}; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> operator/(const vec<dataT, numElements> &lhs,
-                                           const dataT &rhs)
-  { return vec<dataT,numElements>{lhs._impl / rhs}; }
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_VEC(^, t)
 
 
-  // OP is: +=, -=, *=, /=, %=
-  /* When OP is %= available only when: dataT != cl_float && dataT != cl_double
-&& dataT != cl_half. */
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> &operator+=(vec<dataT, numElements> &lhs, 
-                                       const vec<dataT, numElements> &rhs)
-  { lhs._impl += rhs._impl; return lhs; }
+  #define HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_SCALAR(op, t)                       \
+  HIPSYCL_UNIVERSAL_TARGET                                                     \
+  friend vec<t, N> operator op(const vec<t, N> &lhs, const t& rhs) {           \
+    vec<t, N> result;                                                          \
+    for (int i = 0; i < N; ++i) {                                              \
+      result._data[i] = lhs._data[i] op rhs;                                   \
+    }                                                                          \
+    return result;                                                             \
+  }
+
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_SCALAR(%, t)
+
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_SCALAR(+, T)
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_SCALAR(-, T)
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_SCALAR(*, T)
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_SCALAR(/, T)
+
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_SCALAR(&, t)
+
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_SCALAR(|, t)
+
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_BINARY_VEC_OP_VEC_SCALAR(^, t)
+
+  #define HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_VEC(op, t)                         \
+  HIPSYCL_UNIVERSAL_TARGET                                                     \
+  friend vec& operator op(vec& lhs, const vec<t,N>& rhs) {                     \
+    for (int i = 0; i < N; ++i) {                                              \
+      lhs._data[i] op rhs._data[i];                                            \
+    }                                                                          \
+    return lhs;                                                                \
+  }
+
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_VEC(%=, t)
+
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_VEC(+=, T)
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_VEC(-=, T)
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_VEC(*=, T)
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_VEC(/=, T)
+
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_VEC(&=, t)
+
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_VEC(|=, t)
+
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_VEC(^=, t)
+
+
+  #define HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_SCALAR(op, t)                      \
+  HIPSYCL_UNIVERSAL_TARGET                                                     \
+  friend vec& operator op(vec& lhs, const t& rhs) {                            \
+    for (int i = 0; i < N; ++i) {                                              \
+      lhs._data[i] op rhs;                                                     \
+    }                                                                          \
+    return lhs;                                                                \
+  }
+
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_SCALAR(%=, t)
+
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_SCALAR(+=, T)
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_SCALAR(-=, T)
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_SCALAR(*=, T)
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_SCALAR(/=, T)
+
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_SCALAR(&=, t)
+
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_SCALAR(|=, t)
+  
+  template <typename t = T,
+            std::enable_if_t<std::is_integral_v<t>, bool> = true>
+  HIPSYCL_DEFINE_INPLACE_VEC_OP_VEC_SCALAR(^=, t)
+
 
   HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> &operator-=(vec<dataT, numElements> &lhs,
-                                       const vec<dataT, numElements> &rhs)
-  { lhs._impl -= rhs._impl; return lhs; }
+  friend vec& operator++(vec& v) {
+    for(int i = 0; i < N; ++i)
+      ++(v._data[i]);
+    return v;
+  }
 
   HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> &operator*=(vec<dataT, numElements> &lhs,
-                                       const vec<dataT, numElements> &rhs)
-  { lhs._impl *= rhs._impl; return lhs; }
+  friend vec& operator--(vec& v) {
+    for(int i = 0; i < N; ++i)
+      --(v._data[i]);
+    return v;
+  }
 
   HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> &operator/=(vec<dataT, numElements> &lhs,
-                                       const vec<dataT, numElements> &rhs)
-  { lhs._impl /= rhs._impl; return lhs; }
+  friend vec<T,N> operator++(vec& v, int) {
+    vec<T,N> old = v;
 
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> &operator%=(vec<dataT, numElements> &lhs,
-                                       const vec<dataT, numElements> &rhs)
-  { lhs._impl %= rhs._impl; return lhs; }
+    for(int i = 0; i < N; ++i)
+      ++(v._data[i]);
 
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> &operator+=(vec<dataT, numElements> &lhs, const dataT &rhs)
-  { lhs._impl += rhs; return lhs; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> &operator-=(vec<dataT, numElements> &lhs, const dataT &rhs)
-  { lhs._impl -= rhs; return lhs; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> &operator*=(vec<dataT, numElements> &lhs, const dataT &rhs)
-  { lhs._impl *= rhs; return lhs; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> &operator/=(vec<dataT, numElements> &lhs, const dataT &rhs)
-  { lhs._impl /= rhs; return lhs; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> &operator%=(vec<dataT, numElements> &lhs, const dataT &rhs)
-  { lhs._impl %= rhs; return lhs; }
-
-  // OP is: ++, --
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> &operator++(vec<dataT, numElements> &lhs)
-  { lhs += 1; return lhs; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> &operator--(vec<dataT, numElements> &lhs)
-  { lhs -= 1; return lhs; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> operator++(vec<dataT, numElements> &lhs,int)
-  {
-    vec<dataT, numElements> old = lhs;
-    ++lhs;
     return old;
   }
 
   HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> operator--(vec<dataT, numElements> &lhs, int)
-  {
-    vec<dataT, numElements> old = lhs;
-    --lhs; 
+  friend vec<T,N> operator--(vec& v, int) {
+    vec<T,N> old = v;
+
+    for(int i = 0; i < N; ++i)
+      --(v._data[i]);
+
     return old;
   }
 
-  // OP is: &, |, ˆ
-  /* Available only when: dataT != cl_float && dataT != cl_double
-&& dataT != cl_half. */
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
   HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT, numElements> operator&(const vec<dataT,numElements> &lhs,
-                   const vec<dataT,numElements> &rhs) 
-  { return vec<dataT,numElements>{lhs._impl & rhs._impl}; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements> operator|(const vec<dataT, numElements> &lhs, 
-                                            const vec<dataT,numElements> &rhs)
-  { return vec<dataT,numElements>{lhs._impl | rhs._impl}; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements> operator^(const vec<dataT, numElements> &lhs, 
-                                            const vec<dataT,numElements> &rhs)
-  { return vec<dataT,numElements>{lhs._impl ^ rhs._impl}; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements> operator&(const vec<dataT, numElements> &lhs, 
-                                            const dataT &rhs)
-  { return vec<dataT,numElements>{lhs._impl & rhs}; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements> operator|(const vec<dataT, numElements> &lhs, 
-                                            const dataT &rhs)
-  { return vec<dataT,numElements>{lhs._impl | rhs}; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements> operator^(const vec<dataT, numElements> &lhs, 
-                                            const dataT &rhs)
-  { return vec<dataT,numElements>{lhs._impl ^ rhs}; }
-
-  // OP is: &=, |=, ˆ=
-  /* Available only when: dataT != cl_float && dataT != cl_double
-&& dataT != cl_half. */
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements>& operator&=(vec<dataT, numElements> &lhs, 
-                                        const vec<dataT,numElements> &rhs) 
-  { lhs._impl &= rhs._impl; return lhs; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements>& operator|=(vec<dataT, numElements> &lhs, 
-                                        const vec<dataT,numElements> &rhs)
-  { lhs._impl |= rhs._impl; return lhs; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements>& operator^=(const vec<dataT, numElements> &lhs, 
-                                              const vec<dataT,numElements> &rhs)
-  { lhs._impl ^= rhs._impl; return lhs; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements>& operator&=(const vec<dataT, numElements> &lhs, 
-                                              const dataT& rhs)
-  { lhs._impl &= rhs; return lhs; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements>& operator|=(const vec<dataT, numElements> &lhs, 
-                                              const dataT& rhs)
-  { lhs._impl |= rhs; return lhs; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements>& operator^=(const vec<dataT, numElements> &lhs, 
-                                              const dataT& rhs)
-  { lhs._impl ^= rhs; return lhs; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator&& (const vec<dataT, numElements>& lhs, 
-                   const vec<dataT, numElements>& rhs)
-  {
-    using result_type =
-        typename detail::logical_vector_op_result<dataT>::type;
-
-    return vec<result_type, numElements>{
-      lhs._impl && rhs._impl
-    };
+  friend vec<T,N> operator+(vec& v) {
+    return v;
   }
 
   HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator|| (const vec<dataT, numElements> &lhs, 
-                   const vec<dataT, numElements>& rhs)
-  {
-    using result_type =
-        typename detail::logical_vector_op_result<dataT>::type;
-
-    return vec<result_type, numElements>{
-      lhs._impl || rhs._impl
-    };
+  friend vec<T,N> operator-(vec& v) {
+    vec<T,N> result;
+    for(int i = 0; i < N; ++i) {
+      result._data[i] = -(v._data[i]);
+    }
+    return result;
   }
 
-  // OP is: &&, ||
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator&& (const vec<dataT, numElements> &lhs, 
-                          const dataT &rhs)
-  { return lhs && vec<dataT,numElements>{rhs}; }
+  //HIPSYCL_UNIVERSAL_TARGET
+  //friend auto operator &&(const vec& lhs, const vec& rhs) {
 
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator|| (const vec<dataT, numElements> &lhs, 
-                          const dataT &rhs) 
-  { return lhs || vec<dataT,numElements>{rhs}; }
-
-  // OP is: <<, >>
-  /* Available only when: dataT != cl_float && dataT != cl_double
-&& dataT != cl_half. */
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT,numElements> operator <<(const vec<dataT, numElements> &lhs, 
-                                            const vec<dataT, numElements>& rhs) 
-  { return vec<dataT,numElements>{lhs._impl << rhs._impl}; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT,numElements> operator >>(const vec<dataT, numElements> &lhs, 
-                                     const vec<dataT, numElements>& rhs)
-  { return vec<dataT,numElements>{lhs._impl >> rhs._impl}; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT,numElements> operator <<(const vec<dataT, numElements> &lhs, 
-                                            const dataT& rhs)
-  { return vec<dataT,numElements>{lhs._impl << rhs}; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend vec<dataT,numElements> operator >>(const vec<dataT, numElements> &lhs, 
-                                            const dataT& rhs)
-  { return vec<dataT,numElements>{lhs._impl >> rhs}; }
-
-  // OP is: <<=, >>=
-  /* Available only when: dataT != cl_float && dataT != cl_double
-&& dataT != cl_half. */
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements>& operator<<=(const vec<dataT, numElements> &lhs, 
-                                               const vec<dataT, numElements> &rhs)
-  { lhs._impl <<= rhs._impl; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements>& operator>>=(const vec<dataT, numElements> &lhs, 
-                                               const vec<dataT, numElements> &rhs)
-  { lhs._impl >>= rhs._impl; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements>& operator<<=(const vec<dataT, numElements> &lhs, 
-                                               const dataT &rhs)
-  { lhs._impl <<= rhs; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend  vec<dataT, numElements>& operator>>=(const vec<dataT, numElements> &lhs, 
-                                               const dataT &rhs)
-  { lhs._impl >>= rhs; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator== (const vec<dataT, numElements> &lhs,
-                          const vec<dataT, numElements>& rhs)
-  {
-    using result_type =
-        typename detail::logical_vector_op_result<dataT>::type;
-
-    return vec<result_type, numElements>{
-      lhs._impl == rhs._impl
-    };
-  }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator!= (const vec<dataT, numElements>& lhs,
-                          const vec<dataT, numElements>& rhs) 
-  {
-    using result_type =
-        typename detail::logical_vector_op_result<dataT>::type;
-
-    return vec<result_type, numElements>{
-      lhs._impl != rhs._impl
-    };
-  }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator< (const vec<dataT, numElements>& lhs,
-                        const vec<dataT, numElements>& rhs) 
-  {
-    using result_type =
-        typename detail::logical_vector_op_result<dataT>::type;
-
-    return vec<result_type, numElements>{
-      lhs._impl < rhs._impl
-    };
-  }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator> (const vec<dataT, numElements>& lhs,
-                         const vec<dataT, numElements>& rhs)
-  {
-    using result_type =
-        typename detail::logical_vector_op_result<dataT>::type;
-
-    return vec<result_type, numElements>{
-      lhs._impl > rhs._impl
-    };
-  }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator<= (const vec<dataT, numElements>& lhs,
-                          const vec<dataT, numElements>& rhs)
-  {
-    using result_type =
-        typename detail::logical_vector_op_result<dataT>::type;
-
-    return vec<result_type, numElements>{
-      lhs._impl <= rhs._impl
-    };
-  }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator>= (const vec<dataT, numElements>& lhs,
-                          const vec<dataT, numElements>& rhs)
-  {
-    using result_type =
-        typename detail::logical_vector_op_result<dataT>::type;
-
-    return vec<result_type, numElements>{
-      lhs._impl >= rhs._impl
-    };
-  }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator== (const vec<dataT, numElements>& lhs,
-                          const dataT &rhs) 
-  { return lhs == vec<dataT,numElements>{rhs}; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator!= (const vec<dataT, numElements>& lhs,
-                          const dataT &rhs) 
-  { return lhs != vec<dataT,numElements>{rhs}; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator< (const vec<dataT, numElements>& lhs,
-                         const dataT &rhs) 
-  { return lhs < vec<dataT,numElements>{rhs}; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator> (const vec<dataT, numElements>& lhs,
-                         const dataT &rhs) 
-  { return lhs > vec<dataT,numElements>{rhs}; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator<= (const vec<dataT, numElements>& lhs,
-                          const dataT &rhs) 
-  { return lhs <= vec<dataT,numElements>{rhs}; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator>= (const vec<dataT, numElements>& lhs,
-                          const dataT &rhs) 
-  { return lhs >= vec<dataT,numElements>{rhs}; }
-
-  /* Available only when: dataT != cl_float && dataT != cl_double
-&& dataT != cl_half. */
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator~(vec<dataT, numElements>& lhs)
-  { return vec<dataT, numElements>{~lhs._impl}; }
-
-  template<class t = dataT,
-           std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  friend auto operator!(vec<dataT, numElements>& lhs)
-  { return vec<dataT, numElements>{!lhs._impl}; }
-
-  // OP is: +, -, *, /, %
-
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator+ (const dataT &lhs,
-                                            const vec<dataT, numElements> &rhs)
-  { return rhs + lhs; }
-
-
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator* (const dataT &lhs,
-                                            const vec<dataT, numElements> &rhs)
-  { return rhs * lhs; }
-
-
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator- (const dataT &lhs,
-                                            const vec<dataT, numElements> &rhs)
-  { return vec<dataT,numElements>{lhs} - rhs; }
-
-
-
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator/ (const dataT &lhs,
-                                            const vec<dataT, numElements> &rhs)
-  { return vec<dataT,numElements>{lhs} / rhs; }
-
-  template <class t = dataT,
-      std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator% (const dataT &lhs,
-                                            const vec<dataT, numElements> &rhs)
-  { return vec<dataT,numElements>{lhs} % rhs; }
-
-
-
-  // OP is: &, |, ˆ
-  // Available only when: dataT != cl_float && dataT != cl_double && dataT != cl_half.
-    template <class t = dataT,
-      std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator&(const dataT& lhs,
-                                           const vec<dataT, numElements>& rhs)
-  { return rhs & lhs; }
-
-    template <class t = dataT,
-      std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator|(const dataT& lhs,
-                                           const vec<dataT, numElements>& rhs)
-  { return rhs | lhs; }
-
-    template <class t = dataT,
-      std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator^(const dataT& lhs,
-                                           const vec<dataT, numElements>& rhs)
-  { return rhs ^ lhs; }
-
-
-  // OP is: &&, ||
-
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator&&(const dataT& lhs,
-                                            const vec<dataT, numElements>& rhs)
-  { return rhs && lhs; }
-
-
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator||(const dataT& lhs,
-                                            const vec<dataT, numElements>& rhs)
-  { return rhs || lhs; }
-
-
-  // OP is: <<, >>
-  // Available only when: dataT != cl_float && dataT != cl_double && dataT != cl_half.
-    template <class t = dataT,
-      std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator<<(const dataT &lhs,
-                                    const vec<dataT, numElements> &rhs)
-  { return vec<dataT, numElements>{lhs} << rhs; }
-
-    template <class t = dataT,
-      std::enable_if_t<std::is_integral<t>::value>* = nullptr>
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator>>(const dataT &lhs,
-                                    const vec<dataT, numElements> &rhs)
-  { return vec<dataT, numElements>{lhs} >> rhs; }
-
-
-  // OP is: ==, !=, <, >, <=, >=
-
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator==(const dataT& lhs,
-                                            const vec<dataT, numElements>& rhs)
-  { return rhs == lhs; }
-
-
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator!=(const dataT& lhs,
-                                            const vec<dataT, numElements>& rhs)
-  { return rhs != lhs; }
-
-
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator< (const dataT& lhs,
-                                            const vec<dataT, numElements>& rhs)
-  { return rhs >= lhs; }
-
-
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator<=(const dataT& lhs,
-                                            const vec<dataT, numElements>& rhs)
-  { return rhs > lhs; }
-
-
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator> (const dataT& lhs,
-                                            const vec<dataT, numElements>& rhs)
-  { return rhs <= lhs; }
-
-
-  HIPSYCL_UNIVERSAL_TARGET
-  inline
-  friend vec<dataT, numElements> operator>=(const dataT& lhs,
-                                            const vec<dataT, numElements>& rhs)
-  { return rhs < lhs; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  vec<dataT, numElements> &operator=(const vec<dataT, numElements> &rhs)
-  { _impl = rhs._impl; return *this; }
-
-  HIPSYCL_UNIVERSAL_TARGET
-  vec<dataT, numElements> &operator=(const dataT &rhs)
-  {
-    broadcast_scalar_init(typename detail::vector_impl<dataT,numElements>::indices(),
-                          rhs);
-    return *this;
-  }
-
-  template<int Original_vec_size,
-           int... Access_indices>
-  HIPSYCL_UNIVERSAL_TARGET
-  vec<dataT, numElements>& operator=(
-      const detail::vec_swizzle<dataT, Original_vec_size, Access_indices...>& swizzled_vec)
-  {
-    _impl = swizzled_vec.swizzled_access().get();
-    return *this;
-  }
+  //}
 
 private:
-  detail::vector_impl<dataT, numElements> _impl;
+  template <typename Arg>
+  HIPSYCL_UNIVERSAL_TARGET
+  static constexpr int count_num_elements(const Arg &x) {
+    if constexpr(std::is_same_v<Arg, T>)
+      return 1;
+    else if(std::is_same_v<typename Arg::element_type, T>)
+      return x.get_count();
+    // ToDo: Trigger error
+    return 0;
+  }
+
+  template<typename Arg>
+  HIPSYCL_UNIVERSAL_TARGET
+  void partial_initialization(int& current_init_index, const Arg& x) {
+    if constexpr(std::is_same_v<Arg, T>) {
+      _data[current_init_index] = x;
+      ++current_init_index;
+    } else {
+      // Assume we are dealing with another vector
+      constexpr int count = count_num_elements(x);
+      
+      for(int i = 0; i < count; ++i) {
+        _data[i + current_init_index] = x._data[i];
+      }
+      
+      current_init_index += count;
+    }
+  }
+
+  VectorStorage _data;
 };
 
+using char2 = vec<int8_t, 2>;
+using char3 = vec<int8_t, 3>;
+using char4 = vec<int8_t, 4>;
+using char8 = vec<int8_t, 8>;
+using char16 = vec<int8_t, 16>;
 
-#define HIPSYCL_DEFINE_VECTOR_ALIAS(T, alias) \
-  using alias ## 2  = vec<T, 2 >; \
-  using alias ## 3  = vec<T, 3 >; \
-  using alias ## 4  = vec<T, 4 >; \
-  using alias ## 8  = vec<T, 8 >; \
-  using alias ## 16 = vec<T, 16>
+using uchar2 = vec<uint8_t, 2>;
+using uchar3 = vec<uint8_t, 3>;
+using uchar4 = vec<uint8_t, 4>;
+using uchar8 = vec<uint8_t, 8>;
+using uchar16 = vec<uint8_t, 16>;
 
-HIPSYCL_DEFINE_VECTOR_ALIAS(char, char);
-HIPSYCL_DEFINE_VECTOR_ALIAS(short, short);
-HIPSYCL_DEFINE_VECTOR_ALIAS(int, int);
-HIPSYCL_DEFINE_VECTOR_ALIAS(long, long);
-HIPSYCL_DEFINE_VECTOR_ALIAS(float, float);
-HIPSYCL_DEFINE_VECTOR_ALIAS(double, double);
-HIPSYCL_DEFINE_VECTOR_ALIAS(detail::hp_float, half);
+using short2 = vec<int16_t, 2>;
+using short3 = vec<int16_t, 3>;
+using short4 = vec<int16_t, 4>;
+using short8 = vec<int16_t, 8>;
+using short16 = vec<int16_t, 16>;
 
-HIPSYCL_DEFINE_VECTOR_ALIAS(detail::s_char, cl_char);
-HIPSYCL_DEFINE_VECTOR_ALIAS(detail::u_char, cl_uchar);
-HIPSYCL_DEFINE_VECTOR_ALIAS(detail::s_short, cl_short);
-HIPSYCL_DEFINE_VECTOR_ALIAS(detail::u_short, cl_ushort);
-HIPSYCL_DEFINE_VECTOR_ALIAS(detail::s_int, cl_int);
-HIPSYCL_DEFINE_VECTOR_ALIAS(detail::u_int, cl_uint);
-HIPSYCL_DEFINE_VECTOR_ALIAS(detail::s_long, cl_long);
-HIPSYCL_DEFINE_VECTOR_ALIAS(detail::u_long, cl_ulong);
-HIPSYCL_DEFINE_VECTOR_ALIAS(detail::sp_float, cl_float);
-HIPSYCL_DEFINE_VECTOR_ALIAS(detail::dp_float, cl_double);
-HIPSYCL_DEFINE_VECTOR_ALIAS(detail::hp_float, cl_half);
+using ushort2 = vec<uint16_t, 2>;
+using ushort3 = vec<uint16_t, 3>;
+using ushort4 = vec<uint16_t, 4>;
+using ushort8 = vec<uint16_t, 8>;
+using ushort16 = vec<uint16_t, 16>;
 
+using int2 = vec<int32_t, 2>;
+using int3 = vec<int32_t, 3>;
+using int4 = vec<int32_t, 4>;
+using int8 = vec<int32_t, 8>;
+using int16 = vec<int32_t, 16>;
 
-HIPSYCL_DEFINE_VECTOR_ALIAS(signed char, schar);
-HIPSYCL_DEFINE_VECTOR_ALIAS(unsigned char, uchar);
-HIPSYCL_DEFINE_VECTOR_ALIAS(unsigned short, ushort);
-HIPSYCL_DEFINE_VECTOR_ALIAS(unsigned int, uint);
-HIPSYCL_DEFINE_VECTOR_ALIAS(unsigned long, ulong);
-HIPSYCL_DEFINE_VECTOR_ALIAS(long long, longlong);
-HIPSYCL_DEFINE_VECTOR_ALIAS(unsigned long long, ulonglong);
+using uint2 = vec<uint32_t, 2>;
+using uint3 = vec<uint32_t, 3>;
+using uint4 = vec<uint32_t, 4>;
+using uint8 = vec<uint32_t, 8>;
+using uint16 = vec<uint32_t, 16>;
 
+using long2 = vec<int64_t, 2>;
+using long3 = vec<int64_t, 3>;
+using long4 = vec<int64_t, 4>;
+using long8 = vec<int64_t, 8>;
+using long16 = vec<int64_t, 16>;
 
+using ulong2 = vec<uint64_t, 2>;
+using ulong3 = vec<uint64_t, 3>;
+using ulong4 = vec<uint64_t, 4>;
+using ulong8 = vec<uint64_t, 8>;
+using ulong16 = vec<uint64_t, 16>;
 
-namespace detail {
+using float2 = vec<float, 2>;
+using float3 = vec<float, 3>;
+using float4 = vec<float, 4>;
+using float8 = vec<float, 8>;
+using float16 = vec<float, 16>;
 
-template<class T, int N, class Function>
-HIPSYCL_UNIVERSAL_TARGET
-void transform_vector(vec<T,N>& v, Function f)
-{
-  v._impl.transform(f);
+using double2 = vec<double, 2>;
+using double3 = vec<double, 3>;
+using double4 = vec<double, 4>;
+using double8 = vec<double, 8>;
+using double16 = vec<double, 16>;
+
 }
-
-template<class T, int N, class Function>
-HIPSYCL_UNIVERSAL_TARGET
-vec<T,N> binary_vector_operation(const vec<T,N>& a,
-                                 const vec<T,N>& b,
-                                 Function f)
-{
-  return vec<T,N>(a._impl.binary_operation(f, b._impl));
 }
-
-template<class T, int N, class Function>
-HIPSYCL_UNIVERSAL_TARGET
-vec<T,N> trinary_vector_operation(const vec<T,N>& a,
-                                  const vec<T,N>& b,
-                                  const vec<T,N>& c,
-                                  Function f)
-{
-  return vec<T,N>{
-        vector_impl<T,N>::trinary_operation(f, a._impl, b._impl, c._impl)
-  };
-}
-
-template<class T,
-         int Original_vec_size,
-         int... Access_indices>
-HIPSYCL_UNIVERSAL_TARGET
-vec_swizzle<T,Original_vec_size,Access_indices...>&
-vec_swizzle<T,Original_vec_size,Access_indices...>::operator=(
-    const vec<T, vec_swizzle<T,Original_vec_size,Access_indices...>::N>& rhs)
-{
-  _swizzled_access.set(rhs._impl);
-  return *this;
-}
-
-} // namespace detail
-} // namespace sycl
-} // namespace hipsycl
-
 
 #endif
